electron-app/resources/backend/trailing_stop.py
"""
Trailing Stop Logic for Dynamic Stop Loss Management
V2.3.37 FIX: Korrigiert für SQLite (vorher MongoDB-Syntax)
"""

import logging
from typing import Dict, Optional, List

logger = logging.getLogger(__name__)


async def update_trailing_stops(db, current_prices: Dict[str, float], settings):
    """
    Update trailing stops for all open positions
    
    V2.5.1: Trades müssen mindestens 2 Minuten offen sein bevor Trailing Stop aktiviert wird!
    
    Args:
        db: Database manager with trades_db
        current_prices: Dict mapping commodity_id to current price
        settings: Trading settings with trailing stop configuration
    """
    # Trailing Stop ist immer aktiv, Settings werden ignoriert
    trailing_distance = 0.0001  # Sehr enger Abstand, damit Gewinn sofort gesichert wird
    
    try:
        from datetime import datetime, timezone
        
        # V2.3.37 FIX: Korrigiert für SQLite - Hole offene Trades aus der DB
        open_trades = await db.trades_db.get_open_trades() if hasattr(db, 'trades_db') else []
        
        if not open_trades:
            return
        
        updated_count = 0
        for trade in open_trades:
            commodity = trade.get('commodity', 'WTI_CRUDE')
            current_price = current_prices.get(commodity)
            
            if not current_price:
                continue
            
            # V2.5.1: ZEITSCHUTZ - Trade muss mindestens 2 Minuten offen sein
            trade_timestamp = trade.get('timestamp') or trade.get('opened_at')
            if trade_timestamp:
                try:
                    if isinstance(trade_timestamp, str):
                        from dateutil.parser import parse as parse_date
                        opened_at = parse_date(trade_timestamp)
                    else:
                        opened_at = trade_timestamp
                    
                    if opened_at.tzinfo is None:
                        opened_at = opened_at.replace(tzinfo=timezone.utc)
                    
                    age_seconds = (datetime.now(timezone.utc) - opened_at).total_seconds()
                    if age_seconds < 120:  # 2 Minuten Schutz
                        logger.debug(f"⏭️ Trade {trade.get('id')}: Zu jung ({age_seconds:.0f}s < 120s) - Trailing Stop übersprungen")
                        continue
                except Exception as e:
                    logger.debug(f"Zeit-Parse-Fehler: {e}")
            
            trade_type = trade.get('type')
            entry_price = trade.get('entry_price')
            current_stop_loss = trade.get('stop_loss')
            
            if not entry_price:
                continue
            
            new_stop_loss = None
            
            # BUY Trade: Stop-Loss immer auf Einstand oder minimalen Gewinn nachziehen
            if trade_type == 'BUY':
                # Wenn Trade im Gewinn ist, Stop-Loss auf Einstand + minimalen Gewinn
                if current_price > entry_price:
                    potential_stop = max(entry_price * 1.0001, current_price * (1 - trailing_distance))
                    if not current_stop_loss or potential_stop > current_stop_loss:
                        new_stop_loss = round(potential_stop, 2)
            # SELL Trade: Stop-Loss immer auf Einstand oder minimalen Gewinn nachziehen
            elif trade_type == 'SELL':
                if current_price < entry_price:
                    potential_stop = min(entry_price * 0.9999, current_price * (1 + trailing_distance))
                    if not current_stop_loss or potential_stop < current_stop_loss:
                        new_stop_loss = round(potential_stop, 2)
            
            # Update stop loss if changed
            if new_stop_loss and new_stop_loss != current_stop_loss:
                try:
                    # V2.3.37 FIX: Korrigiert für SQLite
                    if hasattr(db, 'trades_db'):
                        await db.trades_db.update_trade_stop_loss(trade['id'], new_stop_loss)
                    updated_count += 1
                    
                    logger.info(
                        f"Trailing Stop updated for {commodity} {trade_type} trade: "
                        f"Stop Loss {current_stop_loss or 'N/A'} -> {new_stop_loss} "
                        f"(Price: {current_price}, Distance: {trailing_distance * 100:.1f}%)"
                    )
                except Exception as update_err:
                    logger.debug(f"Could not update trailing stop: {update_err}")
        
        if updated_count > 0:
            logger.info(f"Updated {updated_count} trailing stops")
    
    except Exception as e:
        logger.error(f"Error updating trailing stops: {e}")


async def check_stop_loss_triggers(db, current_prices: Dict[str, float]) -> List[Dict]:
    """
    Check if any positions should be closed due to stop loss
    
    V2.5.1: Trades müssen mindestens 2 Minuten offen sein bevor SL/TP geprüft wird!
    
    Args:
        db: Database manager with trades_db
        current_prices: Dict mapping commodity_id to current price
    
    Returns:
        List of trade dicts that should be closed
    """
    try:
        from datetime import datetime, timezone
        
        # V2.3.37 FIX: Korrigiert für SQLite - Hole offene Trades aus der DB
        open_trades = await db.trades_db.get_open_trades() if hasattr(db, 'trades_db') else []
        
        if not open_trades:
            return []
        
        trades_to_close = []
        
        for trade in open_trades:
            commodity = trade.get('commodity', 'WTI_CRUDE')
            current_price = current_prices.get(commodity)
            
            if not current_price:
                continue
            
            # V2.5.1: ZEITSCHUTZ - Trade muss mindestens 2 Minuten offen sein
            trade_timestamp = trade.get('timestamp') or trade.get('opened_at')
            if trade_timestamp:
                try:
                    if isinstance(trade_timestamp, str):
                        from dateutil.parser import parse as parse_date
                        opened_at = parse_date(trade_timestamp)
                    else:
                        opened_at = trade_timestamp
                    
                    if opened_at.tzinfo is None:
                        opened_at = opened_at.replace(tzinfo=timezone.utc)
                    
                    age_seconds = (datetime.now(timezone.utc) - opened_at).total_seconds()
                    if age_seconds < 120:  # 2 Minuten Schutz
                        logger.debug(f"⏭️ Trade {trade.get('id')}: Zu jung ({age_seconds:.0f}s < 120s) - SL/TP Check übersprungen")
                        continue
                except Exception as e:
                    logger.debug(f"Zeit-Parse-Fehler: {e}")
            
            trade_type = trade.get('type')
            stop_loss = trade.get('stop_loss')
            take_profit = trade.get('take_profit')
            
            # Check stop loss
            if stop_loss:
                if trade_type == 'BUY' and current_price <= stop_loss:
                    trades_to_close.append({
                        'id': trade['id'],
                        'reason': 'STOP_LOSS',
                        'exit_price': current_price
                    })
                    logger.info(f"Stop Loss triggered for {commodity} BUY: {current_price} <= {stop_loss}")
                
                elif trade_type == 'SELL' and current_price >= stop_loss:
                    trades_to_close.append({
                        'id': trade['id'],
                        'reason': 'STOP_LOSS',
                        'exit_price': current_price
                    })
                    logger.info(f"Stop Loss triggered for {commodity} SELL: {current_price} >= {stop_loss}")
            
            # Check take profit
            if take_profit:
                if trade_type == 'BUY' and current_price >= take_profit:
                    trades_to_close.append({
                        'id': trade['id'],
                        'reason': 'TAKE_PROFIT',
                        'exit_price': current_price
                    })
                    logger.info(f"Take Profit triggered for {commodity} BUY: {current_price} >= {take_profit}")
                
                elif trade_type == 'SELL' and current_price <= take_profit:
                    trades_to_close.append({
                        'id': trade['id'],
                        'reason': 'TAKE_PROFIT',
                        'exit_price': current_price
                    })
                    logger.info(f"Take Profit triggered for {commodity} SELL: {current_price} <= {take_profit}")
        
        return trades_to_close
    
    except Exception as e:
        logger.error(f"Error checking stop loss triggers: {e}")
        return []
