electron-app/resources/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.72.1
anyio==4.11.0
APScheduler==3.11.1
attrs==25.4.0
bcrypt==4.1.3
beautifulsoup4==4.14.2
black==25.9.0
boto3==1.40.67
botocore==1.40.67
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
curl_cffi==0.13.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
fastuuid==0.14.0
feedparser==6.0.12
filelock==3.20.0
flake8==7.3.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core
google-api-python-client
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.49.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
iso8601==2.1.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jq==1.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.79.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
metaapi_cloud_copyfactory_sdk==12.0.0
metaapi_cloud_metastats_sdk==6.0.0
metaapi_cloud_sdk==29.1.0
multidict==6.7.0
multitasking==0.0.12
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.4
oauthlib==3.3.1
openai==1.99.9
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==3.14.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==4.6.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
scipy==1.16.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.37.2
stripe==13.2.0
ta==0.11.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
aiosqlite==0.21.0
openai-whisper==20231117
ffmpeg-python==0.2.0
soundfile==0.12.1
psutil
